Cleanups for various gcc warnings, conditional assignment and type errors
Submitted by: wrowe
Reviewed by: trawick


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/branches/1.4.x@1083815 13f79535-47bb-0310-9956-ffa450edef68

+++ file_io/win32/readwrite.c
@@ -266,7 +266,7 @@ APR_DECLARE(apr_status_t) apr_file_write(apr_file_t *thefile, const void *buf, a
             /* Position file pointer for writing at the offset we are logically reading from */
             apr_off_t offset = thefile->filePtr - thefile->dataRead + thefile->bufpos;
             DWORD offlo = (DWORD)offset;
-            DWORD offhi = (DWORD)(offset >> 32);
+            LONG  offhi = (LONG)(offset >> 32);
             if (offset != thefile->filePtr)
                 SetFilePointer(thefile->filehand, offlo, &offhi, FILE_BEGIN);
             thefile->bufpos = thefile->dataRead = 0;
